Fixed compiler format warnings in cinelerra
Making use of macros defined in inttypes.h

+++ cinelerra/filedv.C
@@ -45,6 +45,7 @@
 #include <unistd.h>
 #include <string.h>
 #include <errno.h>
+#include <inttypes.h>
 
 #include <iostream>
 
@@ -509,7 +510,7 @@ TRACE("FileDV::write_samples 220")
 		stream_lock->lock("FileDV::write_samples 10");
 		if(fseeko(stream, (off_t) audio_frames_written * output_size, SEEK_SET) != 0)
 		{
-			eprintf("Unable to set audio write position to %lli\n", (off_t) audio_frames_written * output_size);
+			eprintf("Unable to set audio write position to %" PRId64 "\n", (int64_t) audio_frames_written * output_size);
 
 			stream_lock->unlock();
 			return 1;
@@ -572,7 +573,7 @@ TRACE("FileDV::write_samples 270")
 		stream_lock->lock("FileDV::write_samples 20");
 		if(fseeko(stream, (off_t) audio_frames_written * output_size, SEEK_SET) != 0)
 		{
-			eprintf("ERROR: Unable to relocate for audio write to %lli\n", (off_t) audio_frames_written * output_size);
+			eprintf("ERROR: Unable to relocate for audio write to %" PRId64 "\n", (int64_t) audio_frames_written * output_size);
 			stream_lock->unlock();
 			return 1;
 		}
@@ -678,7 +679,7 @@ int FileDV::write_frames(VFrame ***frames, int len)
 		stream_lock->lock("FileDV::write_frames");
 		if(fseeko(stream, (off_t) video_position * output_size, SEEK_SET) != 0)
 		{
-			eprintf("Unable to seek file to %lli\n", (off_t)(video_position * output_size));
+			eprintf("Unable to seek file to %" PRId64 "\n", video_position * output_size);
 		}
 		if(fwrite(video_buffer, output_size, 1, stream) < 1)
 		{
@@ -701,7 +702,7 @@ int FileDV::read_compressed_frame(VFrame *buffer)
 
 	if (fseeko(stream, (off_t) video_position * output_size, SEEK_SET))
 	{
-		eprintf("Unable to seek file to %lli\n", (off_t)(video_position * output_size));
+		eprintf("Unable to seek file to %" PRId64 "\n", video_position * output_size);
 	}
 	result = fread(buffer->get_data(), output_size, 1, stream);
 	video_position++;
@@ -718,7 +719,7 @@ int FileDV::write_compressed_frame(VFrame *buffer)
 
 	if (fseeko(stream, (off_t) video_position * output_size, SEEK_SET))
 	{
-		eprintf("Unable to seek file to %lli\n", (off_t)(video_position * output_size));
+		eprintf("Unable to seek file to %" PRId64 "\n", video_position * output_size);
 	}
 	result = fwrite(buffer->get_data(), buffer->get_compressed_size(), 1, stream);
 	video_position++;
@@ -821,7 +822,7 @@ TRACE("FileDV::read_frame 10")
 	stream_lock->lock("FileDV::read_frame");
 	if(fseeko(stream, (off_t) video_position * output_size, SEEK_SET) < 0)
 	{
-		eprintf("Unable to seek file to %lli", (off_t)(video_position * output_size));
+		eprintf("Unable to seek file to %" PRId64, video_position * output_size);
 		stream_lock->unlock();
 		return 1;
 	}
